+++ apps/competition/models.py
@@ -4,7 +4,7 @@ from django.contrib.auth.models import User
 from django.db import models
 from django.db.models import Q
 from django.urls import reverse
-from django.utils.translation import ugettext as _
+from django.utils.translation import ugettext as _, ugettext_lazy as _lazy
 
 from translatable.models import TranslatableModel, get_translation_model
 
@@ -14,10 +14,10 @@ from apps.userprofile.models import Alias, AliasType
 
 class Activity(models.Model):
 
-    title = models.CharField('title', max_length=50)
-    image_url = models.CharField('Image url', max_length=100, blank=True,
-                                 help_text='Use a mirrored image of at least a height of 150px.')
-    desc = models.TextField('description')
+    title = models.CharField(_lazy(u'title'), max_length=50)
+    image_url = models.CharField(_lazy(u'image url'), max_length=100, blank=True,
+                                 help_text=_lazy(u'Use a mirrored image of at least a height of 150px.'))
+    desc = models.TextField(_lazy(u'description'), blank=True)
 
     def __unicode__(self):
         return self.title
@@ -26,60 +26,60 @@ class Activity(models.Model):
         return reverse('activity_details', kwargs={'activity_id': self.id})
 
     class Meta:
+        verbose_name = _lazy(u'activity')
+        verbose_name_plural = _lazy(u'activities')
         ordering = ['title']
-        verbose_name = 'activity'
-        verbose_name_plural = 'activities'
 
 
 class Competition(TranslatableModel):
 
-    STATUS_OPTIONS = ((1, _(u'Open')), (2, _(u'Closed')), (3, _(u'In progress')),
-                      (4, _(u'Finished')))
+    STATUS_OPTIONS = ((1, _lazy(u'Open')), (2, _lazy(u'Closed')), (3, _lazy(u'In progress')),
+                      (4, _lazy(u'Finished')))
 
-    TOURNAMENT_FORMATS = (('single elimination', 'Single elimination'), ('double elimination', 'Double elimination'))
+    TOURNAMENT_FORMATS = (('single elimination', _lazy(u'Single elimination')), ('double elimination', _lazy(u'Double elimination')))
 
     statuses = {
-        1: ['Registration open', 'success'],
-        2: ['Registration closed', 'danger'],
-        3: ['Competition in progress', 'warning'],
-        4: ['Competition finished', 'info'],
+        1: [_lazy(u'Registration open'), 'success'],
+        2: [_lazy(u'Registration closed'), 'danger'],
+        3: [_lazy(u'Competition in progress'), 'warning'],
+        4: [_lazy(u'Competition finished'), 'info'],
     }
 
-    status = models.SmallIntegerField('status', choices=STATUS_OPTIONS)
-    activity = models.ForeignKey(Activity)
-    lan = models.ForeignKey(LAN)
-    challonge_url = models.CharField('Challonge url', max_length=50, blank=True, null=True)
-    team_size = models.IntegerField(default=5, blank=True)
-    start_time = models.DateTimeField(blank=True, null=True)
+    status = models.SmallIntegerField(_lazy(u'status'), choices=STATUS_OPTIONS)
+    activity = models.ForeignKey(Activity, verbose_name=_lazy(u'activity'))
+    lan = models.ForeignKey(LAN, verbose_name=_lazy(u'lan'))
+    challonge_url = models.CharField(_lazy(u'Challonge URL'), max_length=50, blank=True)
+    team_size = models.IntegerField(_lazy(u'team size'), default=5, blank=True)
+    start_time = models.DateTimeField(_lazy(u'start time'), blank=True, null=True)
 
     tournament_format = models.CharField(
-        'Tournament format', max_length=20, blank=True, null=True, choices=TOURNAMENT_FORMATS)
+        _lazy(u'tournament format'), max_length=20, blank=True, choices=TOURNAMENT_FORMATS)
 
     max_participants = models.SmallIntegerField(
-        'Maximum participants', default=0, help_text='The maximum number of participants allowed for a competition.'
-                                                     'Restricts participants based on competition type. 0 means'
-                                                     ' infinite participants are allowed.')
+        _lazy(u'maximum participants'), default=0, help_text=_lazy(u'The maximum number of participants allowed for a competition.'
+                                                                   'Restricts participants based on competition type. 0 means'
+                                                                   ' infinite participants are allowed.'))
 
     use_teams = models.BooleanField(
-        'use teams', default=False, help_text='If checked, participants will be ignored, and will '
-                                              'instead use teams. If left unchecked teams will be ignored, '
-                                              'and participants will be used.')
+        _lazy(u'use teams'), default=False, help_text=_lazy(u'If checked, participants will be ignored, and will '
+                                                            'instead use teams. If left unchecked teams will be ignored, '
+                                                            'and participants will be used.'))
 
     enforce_team_size = models.BooleanField(
-        'enforce teams', default=False, help_text='If checked, teams will require x members (specified in team_size)'
-                                                  ' before being able to sign up.')
+        _lazy(u'enforce teams'), default=False, help_text=_lazy(u'If checked, teams will require x members (specified in team_size)'
+                                                                ' before being able to sign up.'))
 
     enforce_payment = models.BooleanField(
-        'enforce payment', default=False, help_text='If checked, teams will require x members (specified in team_size)'
-                                                    ' with valid tickets before being able to sign up.')
+        _lazy(u'enforce payment'), default=False, help_text=_lazy(u'If checked, teams will require x members (specified in team_size)'
+                                                                  ' with valid tickets before being able to sign up.'))
 
     require_alias = models.BooleanField(
-        'require alias', default=False, help_text='If checked, players will need to register an alias for the '
-                                                  'Activity that the competition belongs to.')
+        _lazy(u'require alias'), default=False, help_text=_lazy(u'If checked, players will need to register an alias for the '
+                                                                'Activity that the competition belongs to.'))
 
     max_match_points = models.SmallIntegerField(
-        'Maximum match points', default=1, help_text='This number represents how many points are needed to win '
-                                                     'a match. E.g. 3 in a BO 5 or 16 in BO 30')
+        _lazy(u'maximum match points'), default=1, help_text=_lazy(u'This number represents how many points are needed to win '
+                                                                   'a match. E.g. 3 in a BO 5 or 16 in BO 30'))
 
     def get_teams(self):
         if self.use_teams:
@@ -153,6 +153,8 @@ class Competition(TranslatableModel):
         return competitions
 
     class Meta:
+        verbose_name = _lazy(u'competition')
+        verbose_name_plural = _lazy(u'competitions')
         ordering = ['status']
         permissions = (
             ('manage', 'Manage compos and matches'),
@@ -160,22 +162,23 @@ class Competition(TranslatableModel):
 
 
 class CompetitionTranslation(get_translation_model(Competition, 'competition')):
-    translated_title = models.CharField('title', max_length=50)
-    translated_description = models.TextField('description',
-                                              help_text='Markdown-enabled. You may also use regular (x)HTML markup. For '
-                                              'blockquotes use the following markup:<br/><br/>&lt;blockquote&gt;<br/>&n'
-                                              'bsp;&nbsp;&nbsp;&nbsp;&lt;p&gt;Quote-text& lt;/p&gt;<br/>&nbsp;&nbsp;&nbsp;&nbsp;&lt;'
-                                              'small&gt;Reference&lt;/small&gt;<br/>&lt;/blockquote&gt;')
+    translated_title = models.CharField(_lazy(u'title'), max_length=50)
+    translated_description = models.TextField(_lazy(u'description'))
 
     def __unicode__(self):
         return self.translated_title
 
+    class Meta:
+        verbose_name = _lazy(u'competition translation')
+        verbose_name_plural = _lazy(u'competition translations')
+
 
 class Participant(models.Model):
-    user = models.ForeignKey(User, null=True)
-    team = models.ForeignKey('team.Team', null=True)
-    competition = models.ForeignKey(Competition)
-    cid = models.CharField('cid', max_length=50, null=True)
+    user = models.ForeignKey(User, verbose_name=_lazy(u'user'), null=True)
+    team = models.ForeignKey('team.Team', verbose_name=_lazy(u'team'), null=True)
+    competition = models.ForeignKey(Competition, verbose_name=_lazy(u'competition'))
+    # Nullable (x3)
+    cid = models.CharField(_lazy(u'cid'), max_length=50, null=True, blank=True)
 
     def __unicode__(self):
         if self.user:
@@ -190,6 +193,8 @@ class Participant(models.Model):
             return True
 
     class Meta:
+        verbose_name = _lazy(u'competition participant')
+        verbose_name_plural = _lazy(u'competition participants')
         unique_together = (
             ('user', 'competition'),
             ('team', 'competition'),
@@ -198,15 +203,16 @@ class Participant(models.Model):
 
 
 class Match(models.Model):
-    matchid = models.CharField('matchid', max_length=50)
-    player1 = models.ForeignKey(Participant, related_name='player1', null=True)
-    player2 = models.ForeignKey(Participant, related_name='player2', null=True)
-    competition = models.ForeignKey(Competition)
-    p1_reg_score = models.CharField('p1_reg_score', max_length=50, null=True)
-    p2_reg_score = models.CharField('p2_reg_score', max_length=50, null=True)
-    final_score = models.CharField('final_score', max_length=50, null=True)
-    state = models.CharField('state', max_length=50)
-    winner = models.ForeignKey(Participant, related_name='winner', null=True)
+    matchid = models.CharField(_lazy(u'match ID'), max_length=50)
+    player1 = models.ForeignKey(Participant, verbose_name=_lazy(u'player 1'), related_name='player1', null=True)
+    player2 = models.ForeignKey(Participant, verbose_name=_lazy(u'player 2'), related_name='player2', null=True)
+    competition = models.ForeignKey(Competition, verbose_name=_lazy(u'competition'))
+    # Nullable (x3)
+    p1_reg_score = models.CharField(_lazy(u'p1 reg score'), max_length=50, null=True, blank=True)
+    p2_reg_score = models.CharField(_lazy(u'p2 reg score'), max_length=50, null=True, blank=True)
+    final_score = models.CharField(_lazy(u'final score'), max_length=50, null=True, blank=True)
+    state = models.CharField(_lazy(u'state'), max_length=50)
+    winner = models.ForeignKey(Participant, verbose_name=_lazy(u'winner'), related_name='winner', null=True)
 
     def get_p1(self):
         if self.player1:
@@ -215,7 +221,7 @@ class Match(models.Model):
             else:
                 return self.player1.user
         else:
-            return 'TBA'
+            return _(u'TBA')
 
     def get_p2(self):
         if self.player2:
@@ -224,7 +230,7 @@ class Match(models.Model):
             else:
                 return self.player2.user
         else:
-            return 'TBA'
+            return _(u'TBA')
 
     def get_compo(self):
         return self.competition.activity.title
@@ -244,3 +250,7 @@ class Match(models.Model):
                     or (user in self.player2.members.all() and player_id == '2'):
                 return True
         return False
+
+    class Meta:
+        verbose_name = _lazy(u'competition match')
+        verbose_name_plural = _lazy(u'competition matches')