+++ apps/lottery/admin.py
@@ -2,13 +2,14 @@
 
 from django.conf import settings
 from django.contrib import admin
+from django.utils.translation import ugettext_lazy as _lazy
 
 from apps.lottery.models import Lottery, LotteryParticipant, LotteryTranslation
 
 
 class LotteryTranslationInlineAdmin(admin.StackedInline):
-    verbose_name = 'Translation'
-    verbose_name_plural = 'Translations'
+    verbose_name = _lazy(u'Translation')
+    verbose_name_plural = _lazy(u'Translations')
     model = LotteryTranslation
     max_num = len(settings.LANGUAGES)
     extra = 1