+++ apps/lottery/models.py
@@ -3,6 +3,7 @@
 from django.contrib.auth.models import User
 from django.db import models
 from django.urls import reverse
+from django.utils.translation import ugettext_lazy as _lazy
 
 from translatable.models import TranslatableModel, get_translation_model
 
@@ -11,8 +12,8 @@ from apps.lan.models import LAN
 
 class Lottery(TranslatableModel):
     lan = models.ForeignKey(LAN)
-    registration_open = models.BooleanField('Open', default=False)
-    multiple_winnings = models.BooleanField('Multiple winnings', default=False, help_text='Allows a user to win more than one price')
+    registration_open = models.BooleanField(_lazy(u'open'), default=False)
+    multiple_winnings = models.BooleanField(_lazy(u'multiple winnings'), default=False, help_text=_lazy(u'Allows a user to win more than one time.'))
 
     def is_participating(self, user):
         for participant in self.lotteryparticipant_set.all():
@@ -32,31 +33,44 @@ class Lottery(TranslatableModel):
         return reverse('details', kwargs={'lottery_id': self.id})
 
     class Meta:
-        verbose_name_plural = 'Lotteries'
+        verbose_name = _lazy(u'lottery')
+        verbose_name_plural = _lazy(u'lotteries')
         permissions = (
             ('draw', 'Can draw winners'),
         )
 
 
 class LotteryTranslation(get_translation_model(Lottery, 'lottery')):
-    title = models.CharField('title', max_length=50)
-    description = models.TextField('description')
+    title = models.CharField(_lazy(u'title'), max_length=50)
+    description = models.TextField(_lazy(u'description'))
 
     def __unicode__(self):
         return self.title
 
+    class Meta:
+        verbose_name = _lazy(u'lottery translation')
+        verbose_name_plural = _lazy(u'lottery translations')
+
 
 class LotteryParticipant(models.Model):
-    lottery = models.ForeignKey(Lottery, editable=False)
-    user = models.ForeignKey(User, editable=False)
+    lottery = models.ForeignKey(Lottery, verbose_name=_lazy(u'lottery'), editable=False)
+    user = models.ForeignKey(User, verbose_name=_lazy(u'user'), editable=False)
 
     def __unicode__(self):
         return unicode(self.user)
+
+    class Meta:
+        verbose_name = _lazy(u'lottery participant')
+        verbose_name_plural = _lazy(u'lottery participants')
 
 
 class LotteryWinner(models.Model):
-    lottery = models.ForeignKey(Lottery, editable=False)
-    user = models.ForeignKey(User, editable=False)
+    lottery = models.ForeignKey(Lottery, verbose_name=_lazy(u'lottery'), editable=False)
+    user = models.ForeignKey(User, verbose_name=_lazy(u'user'), editable=False)
 
     def __unicode__(self):
         return unicode(self.user)
+
+    class Meta:
+        verbose_name = _lazy(u'lottery winner')
+        verbose_name_plural = _lazy(u'lottery winners')