+++ apps/news/admin.py
@@ -2,13 +2,14 @@
 
 from django.conf import settings
 from django.contrib import admin
+from django.utils.translation import ugettext_lazy as _lazy
 
 from apps.news.models import Article, ArticleTranslation
 
 
 class ArticleTranslationInlineAdmin(admin.StackedInline):
-    verbose_name = 'Translation'
-    verbose_name_plural = 'Translations'
+    verbose_name = _lazy(u'Translation')
+    verbose_name_plural = _lazy(u'Translations')
     model = ArticleTranslation
     max_num = len(settings.LANGUAGES)
     extra = 2