+++ apps/payment/admin.py
@@ -1,3 +1,0 @@
-# -*- coding: utf-8 -*-
-
-# Register your models here.