+++ manage/make-translations.sh
@@ -8,4 +8,5 @@ set -eu
 echo "Making translation text file ..."
 $MANAGE makemessages --no-location --locale=nb
 
-echo "Done, now edit it and run manage/compile-translations.sh afterwards."
+echo
+echo "Done. Run manage/setup.sh to compile translations locally."